network/syncer: address some pr comments
